requirements.txt
JSON-log-formatter==1.1.1
asyncio==3.4.3
aiogram==3.17.0
requests==2.32.3
pydantic>=2.0
python-dotenv==1.1.1
